+++ MapboxCoreNavigationTests/Support/DummyURLSessionDataTask.swift
@@ -4,4 +4,8 @@ class DummyURLSessionDataTask: URLSessionDataTask {
     override func resume() {
         //
     }
+    
+    override func cancel() {
+        // 
+    }
 }