+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 * If a reroute request fails, the `Rerouting...` status view update is dismissed. [#1357](https://github.com/mapbox/mapbox-navigation-ios/pull/1357)
 * Completed waypoints now remain on the map but are slightly translucent. [#1364](https://github.com/mapbox/mapbox-navigation-ios/pull/1364)
 * `navigationMapView(_:shapeFor:)` has been renamed to `navigationMapView(_:shapeFor:legIndex:)`. [#1364](https://github.com/mapbox/mapbox-navigation-ios/pull/1364)
+* Renamed methods on `NavigationMapViewDelegate`:    
+    - `NavigationMapViewDelegate.navigationMapView(_:shapeDescribingRoute:)` renamed to `NavigationMapViewDelegate.navigationMapView(_:shapeDescribing:)`.
+    - `NavigationMapViewDelegate.navigationMapView(_:shapeDescribingWaypoints:)` renamed to `NavigationMapViewDelegate.navigationMapView(_:simplifiedShapeDescribingRoute:)`.
+    - `NavigationMapViewDelegate.navigationMapView(_:shapeDescribingWaypoints:)` renamed to `NavigationMapViewDelegate.navigationMapView(_:shapeFor:legIndex:)`.
+    - `NavigationMapViewDelegate.navigationMapView(_:imageForAnnotation:)` renamed to `NavigationMapViewDelegate.navigationMapView(_:imageFor:)`.
+    - `NavigationMapViewDelegate.navigationMapView(_:viewForAnnotation:)` renamed to `NavigationMapViewDelegate.navigationMapView(_:viewFor:)`.
+    [#1338](https://github.com/mapbox/mapbox-navigation-ios/pull/1338)
 
 ## v0.16.2 (April 13, 2018)